chore: move TransactionBuilder to its own file
