+++ packages/transaction-builder/src/index.ts
@@ -1,184 +1,1 @@
-import { Buffer } from 'buffer'
-import {
-  MsgProtoAppStake,
-  MsgProtoAppUnstake,
-  MsgProtoNodeStakeTx,
-  MsgProtoNodeUnjail,
-  MsgProtoNodeUnstake,
-  MsgProtoSend,
-} from './models/msgs'
-import { AbstractProvider } from '@pokt-foundation/pocketjs-abstract-provider'
-import { KeyManager } from '@pokt-foundation/pocketjs-signer'
-import {
-  RawTxRequest,
-  TransactionResponse,
-} from '@pokt-foundation/pocketjs-types'
-import { TxEncoderFactory } from './factory/tx-encoder-factory'
-import { TxMsg, CoinDenom, TxSignature } from './models/'
-
-export type ChainID = 'mainnet' | 'testnet' | 'localnet'
-
-export const DEFAULT_BASE_FEE = '10000'
-
-export class TransactionBuilder {
-  private provider: AbstractProvider
-  private signer: KeyManager
-  private chainID: ChainID
-
-  constructor({
-    provider,
-    signer,
-    chainID = 'mainnet',
-  }: {
-    provider: AbstractProvider
-    signer: KeyManager
-    chainID: ChainID
-  }) {
-    this.provider = provider
-    this.signer = signer
-    this.chainID = chainID
-  }
-
-  /**
-   * Signs and creates a transaction object that can be submitted to the network given the parameters and called upon Msgs.
-   * Will empty the msg list after succesful creation
-   * @param {string} fee - The amount to pay as a fee for executing this transaction, in uPOKT (1 POKT = 1*10^6 uPOKT).
-   * @param {string | undefined} memo - The memo field for this account
-   * @returns {Promise<RawTxResponse | RpcError>} - A Raw transaction Response object or Rpc error.
-   * @memberof TransactionSender
-   */
-  public async createTransaction({
-    fee = DEFAULT_BASE_FEE,
-    memo = '',
-    txMsg,
-  }: {
-    fee?: string | bigint
-    memo: string
-    txMsg: TxMsg
-  }): Promise<RawTxRequest> {
-    // Let's make sure txMsg is defined.
-    if (!txMsg) {
-      throw new Error('txMsg should be defined.')
-    }
-
-    const entropy = Number(
-      BigInt(Math.floor(Math.random() * Number.MAX_SAFE_INTEGER)).toString()
-    ).toString()
-
-    const signer = TxEncoderFactory.createEncoder(
-      entropy,
-      this.chainID,
-      txMsg,
-      fee.toString(),
-      CoinDenom.Upokt,
-      memo
-    )
-
-    const bytesToSign = signer.marshalStdSignDoc()
-
-    const txBytes = await this.signer.sign(bytesToSign.toString('hex'))
-
-    const marshalledTx = new TxSignature(
-      Buffer.from(this.signer.getPublicKey(), 'hex'),
-      Buffer.from(txBytes, 'hex')
-    )
-
-    const rawHexBytes = signer.marshalStdTx(marshalledTx).toString('hex')
-
-    return new RawTxRequest(this.signer.getAddress(), rawHexBytes)
-  }
-
-  public async submit({
-    fee = DEFAULT_BASE_FEE,
-    memo = '',
-    txMsg,
-  }: {
-    fee?: string | bigint
-    memo: string
-    txMsg: TxMsg
-  }): Promise<TransactionResponse> {
-    const tx = await this.createTransaction({ fee, memo, txMsg })
-
-    return await this.provider.sendTransaction(tx)
-  }
-
-  public async submitRawTransaction(
-    tx: RawTxRequest
-  ): Promise<TransactionResponse> {
-    return await this.provider.sendTransaction(tx)
-  }
-
-  /**
-   * Adds a MsgSend TxMsg for this transaction
-   * @param {string} fromAddress - Origin address
-   * @param {string} toAddress - Destination address
-   * @param {string} amount - Amount to be sent, needs to be a valid number greater than 0
-   * @returns {MsgProtoSend} - The unsigned Send message.
-   */
-  public send(
-    fromAddress: string,
-    toAddress: string,
-    amount: string
-  ): MsgProtoSend {
-    return new MsgProtoSend(fromAddress, toAddress, amount)
-  }
-
-  /**
-   * Adds a MsgAppStake TxMsg for this transaction
-   * @param {string} appPubKey - Application Public Key
-   * @param {string[]} chains - Network identifier list to be requested by this app
-   * @param {string} amount - the amount to stake, must be greater than 0
-   * @returns {MsgProtoAppStake} - The unsigned App Stake message.
-   */
-  public appStake(
-    appPubKey: string,
-    chains: string[],
-    amount: string
-  ): MsgProtoAppStake {
-    return new MsgProtoAppStake(appPubKey, chains, amount)
-  }
-
-  /**
-   * Adds a MsgBeginAppUnstake TxMsg for this transaction
-   * @param {string} address - Address of the Application to unstake for
-   * @returns {MsgProtoAppUnstake} - The unsigned App Unstake message.
-   */
-  public appUnstake(address: string): MsgProtoAppUnstake {
-    return new MsgProtoAppUnstake(address)
-  }
-
-  /**
-   * Adds a NodeStake TxMsg for this transaction
-   * @param {string} nodePubKey - Node Public key
-   * @param {string[]} chains - Network identifier list to be serviced by this node
-   * @param {string} amount - the amount to stake, must be greater than 0
-   * @param {URL} serviceURL - Node service url
-   * @returns {MsgProtoNodeStakeTx} - The unsigned Node Stake message.
-   */
-  public nodeStake(
-    nodePubKey: string,
-    chains: string[],
-    amount: string,
-    serviceURL: URL
-  ): MsgProtoNodeStakeTx {
-    return new MsgProtoNodeStakeTx(nodePubKey, chains, amount, serviceURL)
-  }
-
-  /**
-   * Adds a MsgBeginUnstake TxMsg for this transaction
-   * @param {string} address - Address of the Node to unstake for
-   * @returns {MsgProtoNodeUnstake} - The unsigned Node Unstake message.
-   */
-  public nodeUnstake(address: string): MsgProtoNodeUnstake {
-    return new MsgProtoNodeUnstake(address)
-  }
-
-  /**
-   * Adds a MsgUnjail TxMsg for this transaction
-   * @param {string} address - Address of the Node to unjail
-   * @returns {MsgProtoNodeUnjail} - The unsigned Node Unjail message.
-   */
-  public nodeUnjail(address: string): MsgProtoNodeUnjail {
-    return new MsgProtoNodeUnjail(address)
-  }
-}
+export * from './tx-builder'
